requirements.txt
boto3==1.9.248
botocore==1.12.248
Click==7.0
docutils==0.15.2
jmespath==0.9.4
patientpop-python-utils==0.1
pkg-resources==0.0.0
pycryptodome==3.9.0
python-dateutil==2.8.0
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.6
